# Simplify compressed length code