This is #519 by @rolandshoemaker with different code.

There is no difference in the speed of `msg.Len` with compression, but there is a noticeable speed up without compression.

I have no idea if this is worth merging or not, but I could see some obvious inefficiencies in that PR and I decided to address them.

```
name                      old time/op  new time/op  delta
MsgLength-8               1.65µs ± 6%  1.65µs ± 6%     ~     (p=0.968 n=10+9)
MsgLengthNoCompression-8   132ns ± 6%    15ns ± 3%  -88.61%  (p=0.000 n=10+10)
```

Original pull request (#519): "Removes a bunch of duplication in the msg.Len path ~and exposes the method used to calculate the compressed length of a RR given a provided compression map~."